refactor: extract io buffer and allocator to io mod
