+++ foyer-storage/src/io/bytes.rs
